pyproject.toml
[tool.poetry]
name = "evol-aie4"
version = "0.1.0"
description = ""
authors = ["Akash Shetty <akash.shetty@iicontent.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.12,<3.13"
requests = "^2.32.3"
python-dotenv = "^1.0.1"
langchain = "^0.3.0"
langchain-core = "^0.3.1"
langchain-openai = "^0.2.0"
langgraph = "^0.2.22"
langchain-community = "^0.3.0"
faiss-cpu = "^1.8.0.post1"
qdrant-client = "^1.11.2"
pandas = "^2.2.2"
pymupdf = "^1.24.10"

[tool.poetry.group.dev.dependencies]
pytest = "^8.3.2"
python-dotenv = "^1.0.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
